requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.3  # Cambiar a Flask>=2.3.3,<2.4.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
PyJWT==2.8.0
python-dotenv==1.1.0
Werkzeug==2.3.7  # Cambiar a esta versión compatible
